refactor: use direct service imports in frontend
