+++ app/frontend/src/composables/history/useGenerationHistory.ts
@@ -2,7 +2,7 @@ import { computed, ref, unref, type ComputedRef } from 'vue';
 import type { MaybeRefOrGetter } from 'vue';
 
 import { debounce, type DebouncedFunction } from '@/utils/async';
-import { listResults as listHistoryResults } from '@/services';
+import { listResults as listHistoryResults } from '@/services/history/historyService';
 import type {
   GenerationHistoryQuery,
   GenerationHistoryResult,
